Enable use of `tput` to auto align printed messages
This allows users of various size terminals to have elegantly printed
messages.  This also allows users to adjust the size of their terminals
while the script is running and have new lines auto-adjust to fit the
new size.

In addition, filenames that are too long to be displayed are now culled
with an ellipsis (…) at the end.