Implemented progress percentage on each task!
This is done using a FIFO pipe.  The FIFO is created in /tmp and is
removed upon script completion (success or no).  The information stored
is held in RAM (good consquence of named pipes) so writing to disk is
essentially nonexistent.

There is a bunch of code cleanups as well, removing the FIND_FLACS from
each task -- now is only done once at the beginning of the script.  This
is also combined with the progress percentage.

More TODO's as well.