Removed progress percentage code
This proved to be inconsistent using FIFO pipes as multiple process
caused seemingly random hanging (orphaned pipes, or lockout).  I imagine
this could be accomplished using System V IPC, rather than FIFO/Unix
Sockets.

The biggest problem is maintaining atomicity when grabbing the number of
FLACs completed so far.  If running multiple cores (6), you'll have 6
processes/threads that will be vying for information at the same time at
some point, creating a race condition.  FIFO pipes maintain they're
atomic as long as the data going into them are smaller or equal to
PIPE_BUF.

Maintaining that atomicity required blocking FIFOs and that effectively
cut the parallel processes down to 1 or 2 AND it created time wasted by
waiting for the FIFO to be clear of writing/reading by other processes.

It might be worth looking into at a future date, but the benefit is
cosmetic, and the previous attempts hindered the scripts effectivenes.