Added --replaygain, -g argument, minor fixes
Replaygain is now calculated for album and track gain.  With this
option, replaygain is added via VORBIS_TAGS.  There are 5 tags that are
added to each FLAC file and will require that users keep the 5 tags in
the USER CONFIGURATION in order to keep the created added replaygain.

The replaygain is applied on a directory above the FLAC files the script
finds (ie the directory holding the FLAC files is the album directory).
This allows both the album and track replaygain tags to be added.

Documentation about the new --replaygain, -g option is forthcoming.

I still need to find a way to add the replaygain tags without relying on
using a directory.

Various fixes have been applied as well.  Adding ALBUMARTIST tag now
allows you to correctly re-tag the files even if the ALBUMARTIST
variable is written as 'album_artist' or 'album artist'.  This is case-insensitve.

Fixes have been applied to make sure the correct VORBIS_TAGS are grabbed
from the FLAC files.

Code cleanup.