Cleaned up functions that print current operation
Each function now removes the ${FALLBACK} check on every
call.  This should minutely speed things up as the check
for 'tput' is only invoked once upon startup.

Reading each function should be easier as the fallback
is separated from the dynamic functions.  This will also
result in a minute speedup in execution as each function
is smaller/simpler to load.