Removed dependence on 'grep'
This essentially removes all the external commands
that are not 'needed' to execute successfully having
been replaced with pure BASH solutions.

The only leftover command is 'sleep' which CAN be a
BASH builtin (through loadables), but is also an
external command (from 'coreutils').

Since it doesn't seem common to have 'sleep' builtin,
removing the explicit check for it seems silly at this
point.