Removed fallback printing
Since 'tput' is no longer and 'stty' is a required dependency,
the fallback printing of current operations is no longer
required.  This will make debugging in the future a HUGE
blessing!

In short, this means ALL users now have the ability to see
dynamically printed messages (in relation to terminal width)!

Added check for 'stty' presence on systems