Finally fixed issue with case-insensitive tags
Even though metaflac follows the specification of allowing
either upper or lower case TAG FIELDS, the user may have
tag FIELDS that are a combination of both.  This fix
ensures all fields are processed, regardless of the case
the user specifies under USER CONFIGURATION.

Optimized the removal of TAG FIELD key using shell wildcard,
which also aids in creating case-insensitivity.

Because of the wildcards, this allows the script to NOT
require BASH 4 or greater!