Automatically check for number of CORES to use
Now, this script checks /proc/cpuinfo for the number
of useable cores that may be present.  If /proc is
mounted and /proc/cpuinfo exists, this will be used,
otherwise the script will default to 2 cores.  The
user has the option to specify how many cores he/she
wants by uncommenting the CORES option and specifying
the number of cores wanted.