Fixed stray process on SIGINT when using auCDtect
The switch from a straight VARIABLE to using BASH's 'mapfile'
into an array, led stray processes when sending SIGINT to the
script.  This only happened when calling upon auCDtect and grabbing
its output.

Now, the old behavior is restored and 'mapfile' is used for post
processing the created array.  This still allows us to get away
without using 'tail'.

Small enhancements/fixes to the summary output as well.