Fixed nasty bug with tags removing spaces on retag
This bug was introduced on the transition from
'echo' to 'printf'.  Due to how 'eval' processes
its arguments, using a straight 'printf' drop-in
replacement for echo yielded a small bug due to the
fact that 'printf' requires an extra argument before
the string to be printed is called.

By escaping some quotes around the string (to be printed)
we now can have the same behavior as before.

A really nasty little bug and another reason why
'eval' is so dangerous!  It is worth looking into ways
of processing variables within declared variables without
'eval' in the near future.

The amount of users affected by this bug is _probably_ very
small, but nonetheless, redoflacs should be updated
immediately.