Rewrote much of the positional parameter handling
This script now makes use of BASH's builtin 'getopts'
program.  This allows for greater flexiblity in calling
arguments about script invocation.

The only addition 'getopts' brings is allowing the short
style arguments to be concatenated on invocation:

	Instead of doing this:
	redoflacs -c -r -p -m directory_of_flacs

	We can do this:
	redoflacs -crpm directory_of_flacs

This can be combined with or without the concatenation as
well as long style arguments being in the mix:

	redoflacs -crp -a -G --md5check directory_of_flacs

Code cleanup and bug fixes as well!