Lots of optimizations and code cleanup!
The way this was done was to minimize the call to
metaflac on each operation.  We now call it once
(or twice depending) with all the files/options
wanted, storing what we want in array.

Whenever we want some data, we call up the array
and not another external call.  This can greatly
speed up execution times.

Functions in which the optimizations were applied:
  - ReplayGain optimizations
	* This operation has an average speed
	  increase of 12%!

  - Compress optimizations
	* Slightly faster!

  - Redo Tags optimizations
    * This operation can be up to 45% faster
	  than the previous version, averaging
	  37% speed increase!

Cleaned up how the script interprets newlines inside
the printing functions

The terminal cursor is now hidden upon script execution,
providing a cleaner, more professional interface.  The
cursor is restored upon script exit