Find the width of a filename instead of length
This alleviates the issues with multibyte characters not being
interpreted correctly.  Before, the length of a filename was
determined by how many characters there were.  Unfortunately, if
a filename had multibyte characters, this would not match up with
the number of blocks a filename took up in a given terminal.

Now we use 'wc' (part of Coreutils) to find out how many 'blocks'
a filename takes up, fixing this issue