Greatly speed up auCDtect operations
This is accomplished by running through each FLAC file
with auCDtect on it's fastest operation (-m40).  If any file
is not 100% verifiable CDDA, the file will run through auCDtect
once again on it's slowest operation (-m0).  This will make it
so needless files that are CDDA quality are not run through the
slowest analyzer needlessly