Fix spinner blinking on fast operations
[ FIXED ]  Each file that's printed to STDOUT is now done by skipping
the first 3 characters, so spaces are not written to the screen.  This
is done by literally positioning the cursor 3 characters after the
farthest left side of the terminal.  This allows the spinner to be
fluid no matter which operation is running

[ FIXED ]  Removed antiquated skipped message when compressing FLAC
files that have already been compressed at the level specified