Fix broken regex for number of jobs ending in '0'
Previously the regular expression omitted '0' as a valid integer, when
determining the number of jobs to run.  If the number of jobs to run
ended in zero, the regular expression would fail, making '10' jobs for
example be interpreted as zero jobs.

This fix uses the [:digit:] character class and does an explicit check
for '0' as the number of jobs specified.