Only show completely finished items on SIGINT
When a user invokes Control-C/SIGINT, only show the number of items
successfully completed, not including the items that were interrupted.

This is accomplished by subtracting the current number of items
processed via $iteration by the number of running children processes.

This only occurs during an interruption to an operation.