Change function names from '__foo()' to '_foo()'
The original logic for prepending function names with two undersocres
no longer applies.  Instead, the switch from two underscores to one,
seems to be most widely accepted as a way to indicate locality.

This script will use that convention.  A single underscore details
locality to the script itself, not local to a function or statement.

_aucdtect() was renamed to _aucdect_cmd() to not interfere with this
new change.

_sox() was renamed to _sox_cmd() to also not interfere with this new
change.