Update configuration file
With the changes to redoflacs since the last release, this commit fixes
a variety of spelling errors and cleans up the format somewhat.  The
structure has been slightly changed, as well as the REVISION number
format at the bottom of the configuration.

The way the configuration is generated is a bit more sane, created by a
hereodoc.  Escaping of special characters (double quotes, backticks,
etc) is no longer required and should make it easier to maintain.

There are also minor tweaks in determining the script revision and
configuration revision number, and as such, the configuration revision
number has been incremented.