refactor: change progress output
This is the beginning stages of cleaning up progress output.  The idea
is to radically simplify the logic for displaying the current progress
of an operation.

A benefit is the intended removal of 'wc' as a dependency, which will
greatly speed up this program as 'wc' is called for _EVERY_ single item
to be processed _AND_ during the progress of each operation.

While a bigger commit, this change is functional where the first stages
of the refactor is reflected as:

  * _test() function as the guinea pig for changes to progress output
  * Removal of 'wc' usage
  * Simplification of displaying item and operational progress

While numerous functions were touched, these are foundational changes
which build the skeleton for future changes.

_test():
  First stage of rewrite, removing the progress bar, keeping the
  percentage completed during operation, and simplifying the calls to
  various functions.

_run_parallel():
  Updated to support the progress output changes. Simplified some of the
  code, and removed the progress/filename output variables.

_print_ok():
_print_fail():
_print_check():
_printf_fail():
  Broke out much of the work done in _print_item() to separate functions
  allowing easier understanding of context when viewing code which makes
  use of these functions.

_print_item():
  Made into a stub function, which currently just displays the currently
  processed item with formatting to ensure it fits within the terminal
  width or 79 characters if the width of the terminal is larger.

  Currently, this only handles 79 characters as the maximum width.  If
  terminal windows are narrower, this function fails.

_truncate_filename():
  Usage removed from _test().  The intent is to eventually remove this
  function.