feat: add context to each operation title
Instead of printing a given message for an operation with issue tracking
at the end of a given operation (across ALL files), now the operation
title message is updated on each failure, per file to show the total
number of issues as files are processed.

Additionally, if a user sends SIGINT, this is also captured (like
before) but is displayed on the operation title message as a label.