feat: dynamically add operation title status
Instead of handling interruption, all operation titles will now handle
whether they completed or did not run, while still displaying the number
of issues, if any.

Additionally, the logic to determine which operation to run, in which
order has been simplified and is directly added to an array in the order
the operations should be run.

The summary of operations has been cleared, pending possible rewrite or
removal.