feat: remove system config and countdown
This script will no longer support running as root via system files.
If a user runs as root, a root-owned config file is placed in root's
$HOME directory.

This simplifies management of the configuration file, in addition to
no longer having to support configuration versions.

It's recommended to create a new configuration file (won't clobber) for
each released version of redoflacs and perform a diff to see if anything
is worth adding/changing.