feat: use preserve modtime for 'flac'/'metaflac'
This breaks up the all usage of 'metaflac' and 'flac' to include the
option to preserve modification times (which is now the default in the
configuration file).

Additionally, the parsing of extra options are now in arrays to be make
it easy to extended with additional global defaults in the future.