feat: simplify trap handling and manage log notice
Instead of trapping on SIGINT and EXIT in a variety of cases, now, the
traps are done in the 'main' portion of script only.

All operation-specific traps are removed and jobs are still taken care
of with cleanup.

As a result, the '_trap_sigint()' fuction is now '_cleanup()' where
there is special handling if the signal received is SIGINT versus EXIT.

This simplifies and the cleanup needed in a variety of areas as it is
centralized to one place now.

The log messaging to the user is simplified and called on cleanup only,
making management of it clearer.

Finally, line wrapping is now dynamically determined and only used when
printing an item during an operation. This allows users with line
wrapping already disabled to not have the terminal enable line wrapping
after the script has run.