feat: add support for '--remove-all-tags-except'
This is a safer operation than removing all tags and applying the ones
to keep, instead only the tags to remove, are removed.

This has the same effect as the older way where multiple tag fields of
the same type are supported.

This is only for users with metaflac >= 1.4.3, those with older versions
still have retagging done the old way.