feat: allow user-defined WAV file directory
Users now have the ability to specify a directory to be used for storing
WAV files which are temporarily created when using operations which need
to have WAV files instead of operating on FLAC files, directly.

This is managed by a configuration option, of which, users will need to
generate a new configuration file in order for the changes to take
effect.

Additionally, all 'remove' commands are safeguarded from operating on
null or possible problematic pathnames if the variables in those
commands somehow are undefined or unset.