feat: improve artwork/spectrogram locations
Previously, there was a possibility where two or more files could have
the same name, but in different directories.

This would cause a possible clobbering of FLAC artwork and spectrograms
if using a user-chosen directory.

Now, the same logic is used as temporary WAV files are stored by using
the, already unique, directory hierarchy where the FLAC files are stored
as a subdirectory to the user-chosen directory.

This ensures no clobbering can take place, even with many multiple jobs
running at the same time.

Additionally, the configuration file has been updated to reflect this
change in pathnames.