fix: ensure SoX reports on failure
Previously, reading SoX's output would do so over a large number of
chunks (1000 characters), which would fail as the delimiter would never
be honored when SoX receives a failure (eg: broken file).

Now, all output is broken up by spaces and parsed, looking for the
values we want (percentage) and all other output is stored, and error
logging is done if the output stored is a SoX error.

Additionally, this also addresses a possible concern that if there was a
very large set of FLAC tags (VorbisComment), the previous 'read' using
1000 characters, would not be enough and the resultant captured output
would break the progress percentage checks.

The script would still function, but the output would be broken with
interspersed metadata.