feat: check for writable user-defined directories
Additionally, simplify the similar code for each directory by looping
over all configuration keys which do directory management and check for
existence and if writable.

Finally, the 'error_log' configuration value is empty, by default, which
points to the user's $HOME directory if left blank, to match the logic
for all other configuration directory-based options.

As a result, the generated configuration was updated to reflect this and
was re-ordered to place the directory-based options at the bottom of the
configuration.