fix: use filename to anchor error messages
This is a very subtle bug that was not caught as the evaluation of the
conditional with the $base variable would STILL succeed as $base was
empty and the since the 'read' command split on whitespace, the glob
would match the ending colon for the previous 'read'.

While this may be "good enough", using part of the item's base filename
as an additional anchor ensures we capture the "correct" ERROR message
from 'flac'.