fix: ensure exit code is maintained on interrupt
Instead of relying on functions which are use by the _cleanup() function
via traps, instead, store a global exit code to return on script exit.

Since the _cleanup() function may be called twice (interruption and
EXIT), this ensures the trap can exit the shell with that global exit
code value.