feat: simplify cleanup and signal handling
Instead of having the _cleanup() function called twice via interrruption
and EXIT traps, the interruption traps set the global exit code and the
final EXIT trap calls the _cleanup() function to perform cleanup if the
user interrupted the script and safely ends the script, as it would
normally.

This simplifies having to check for abnormal and EXIT signals, and
instead performs action on abnormal signals and carries on with
additional cleanup.

Additionally, these changes make the cleanup handling atomic as the EXIT
trap can only be executed once, meaning a SIGINT followed by SIGTERM
will not cause collision as the _cleanup() function is run once during
EXIT only.