fix: allow variables in config to interpolate
The work to use 'export' over 'eval' made variables included as part of
the configuration option value to no longer expand.

Instead of just using 'eval', instead, the value of the configuration is
now checked for a variable and evaluated/interpolated.

Additionally, command substitution is checked before trying to evaluate
the configuration value and if detected, the literal value is used
instead to prevent code injection.